llama

+++ Common/src/main/java/strategy/MarketDeployStrategy.java
@@ -18,7 +18,7 @@ public interface MarketDeployStrategy {
      * @param GuaranteeRatio 产品保本率
      * @return
      */
-    public CPPIMarketDeploy DefaultCPPIDeploy(int PortValue, double RiskMulti, double GuaranteeRatio) throws RemoteException, NotInitialedException, MWException;
+    public CPPIMarketDeploy DefaultCPPIDeploy(double PortValue, double RiskMulti, double GuaranteeRatio) throws RemoteException, NotInitialedException, MWException;
 
     /**
      * 回测区间的CPPI策略
@@ -29,7 +29,7 @@ public interface MarketDeployStrategy {
      * @param endDate
      * @return
      */
-    public CPPIMarketDeploy CustomizedCPPIDeploy(int PortValue, double RiskMulti, double GuaranteeRatio, String startDate, String endDate) throws RemoteException, NotInitialedException, MWException;
+    public CPPIMarketDeploy CustomizedCPPIDeploy(double PortValue, double RiskMulti, double GuaranteeRatio, String startDate, String endDate) throws RemoteException, NotInitialedException, MWException;
 
     /**
      * 当前日期生成的风险平价策略
